# volumes/catalog_tool_ngsi/swagger_server/models/entity.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.attribute import Attribute  # noqa: F401,E501
from swagger_server import util


class Entity(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, attribute_name: Attribute=None):  # noqa: E501
        """Entity - a model defined in Swagger

        :param attribute_name: The attribute_name of this Entity.  # noqa: E501
        :type attribute_name: Attribute
        """
        self.swagger_types = {
            'attribute_name': Attribute
        }

        self.attribute_map = {
            'attribute_name': 'AttributeName'
        }
        self._attribute_name = attribute_name

    @classmethod
    def from_dict(cls, dikt) -> 'Entity':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Entity of this Entity.  # noqa: E501
        :rtype: Entity
        """
        return util.deserialize_model(dikt, cls)

    @property
    def attribute_name(self) -> Attribute:
        """Gets the attribute_name of this Entity.


        :return: The attribute_name of this Entity.
        :rtype: Attribute
        """
        return self._attribute_name

    @attribute_name.setter
    def attribute_name(self, attribute_name: Attribute):
        """Sets the attribute_name of this Entity.


        :param attribute_name: The attribute_name of this Entity.
        :type attribute_name: Attribute
        """

        self._attribute_name = attribute_name